refactor(cli): reuse commands
